+++ components/identifier-generator/back/tests/Specification/Domain/Model/IdentifierGeneratorIdSpec.php
@@ -13,22 +13,22 @@ use PhpSpec\ObjectBehavior;
  */
 class IdentifierGeneratorIdSpec extends ObjectBehavior
 {
-    function let()
+    public function let()
     {
         $this->beConstructedThrough('fromString', ['2038e1c9-68ff-4833-b06f-01e42d206002']);
     }
 
-    function it_is_an_identifier_generator_id()
+    public function it_is_an_identifier_generator_id()
     {
         $this->shouldBeAnInstanceOf(IdentifierGeneratorId::class);
     }
 
-    function it_returns_an_identifier_generator_id()
+    public function it_returns_an_identifier_generator_id()
     {
         $this->asString()->shouldReturn('2038e1c9-68ff-4833-b06f-01e42d206002');
     }
 
-    function it_cannot_be_instantiated_with_not_uuid()
+    public function it_cannot_be_instantiated_with_not_uuid()
     {
         $this->beConstructedThrough('fromString', ['not_uuid']);
         $this->shouldThrow(\InvalidArgumentException::class)->duringInstantiation();