+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/job/product/edit/content/structure/attributes-selector.js
@@ -114,6 +114,7 @@ define([
       });
 
       this.trigger('selected:update:after', this.selected);
+      this.updateAttributeList();
     },
 
     /**